feat: Дописал Бэк, удалил старые файлы

+++ backend/src/controller/task.controller.sequelize.ts
@@ -4,51 +4,86 @@ import { Categories } from '../models/categories.model.js';
 
 export const createTask: RequestHandler = async (req: any, res: any, next: NextFunction) => {
   try {
-    const tasks = await Tasks.create({ ...req.body });
-    res.json({ message: 'makes cool', data: tasks });
+    const category = await Categories.findOrCreate({
+      where: { id: 999, categories_name: 'Something' },
+    });
+
+    const tasks = await Tasks.create({
+      ...req.body,
+    });
+
+    await tasks.addCategory(999);
+
+    res.json({ data: tasks, category });
   } catch (error) {
     next(error);
-    res.status(400).json({ msg: 'Create Task error' });
+    res.status(400).json({ msg: 'Create Task error ' + error });
   }
 };
 
 export const getAllTask: RequestHandler = async (req: any, res: any, next: NextFunction) => {
   try {
-    const allTasks: Tasks[] = await Tasks.findAll({
-      include: { model: Categories, attributes: ['categories_name'] },
-    });
+    const allTasks: Tasks[] = await Tasks.findAll({ include: { model: Categories } });
     res.json({ message: 'makes cool', data: allTasks });
   } catch (error) {
     next(error);
-    res.status(400).json({ msg: 'Get all Task error' });
-  }
-};
-export const updateTask: RequestHandler = async (req: any, res: any, next: NextFunction) => {
-  try {
-    const { id } = req.params;
-    await Tasks.update({ ...req.body }, { where: { id } });
-    const tasks: Tasks | null = await Tasks.findByPk(id);
-    res.json({ data: tasks });
-  } catch (error) {
-    next(error);
-    res.status(400).json({ msg: 'Update Task error' });
+    res.status(400).json({ msg: 'Get all Task error ' + error });
   }
 };
 
-//Сделать каскадное удаление
+export const updateTask: RequestHandler = async (req: any, res: any, next: NextFunction) => {
+  try {
+    const { id } = req.params;
+    if (!id) {
+      return res.status(404).json({ error: 'Task not found' });
+    }
+    await Tasks.update({ ...req.body }, { where: { id } });
+    const tasks: Tasks | null = await Tasks.findByPk(id, {
+      include: { model: Categories, attributes: ['categories_name'] },
+    });
+    res.json({ data: tasks });
+  } catch (error) {
+    next(error);
+    res.status(400).json({ msg: 'Update Task error ' + error });
+  }
+};
+
 export const deleteTask: RequestHandler = async (req: any, res: any, next: NextFunction) => {
   try {
     const { id } = req.params;
     const deletedTask: Tasks | null = await Tasks.findByPk(id);
+    if (!deletedTask) {
+      return res.status(404).json({ error: 'Task not found' });
+    }
 
     await Tasks.destroy({ where: { id } });
 
-    //Ниже попытка сделать удаление каскадным
-    // const user:Tasks | null = await Tasks.findOne({where: {id:}});
-    //  await user.destroy();
     res.json({ data: deletedTask });
   } catch (error) {
     next(error);
     res.status(400).json({ msg: 'Delete Task error ' + error });
+  }
+};
+
+export const removeCategoryFromTask = async (req: any, res: any, next: NextFunction) => {
+  try {
+    const taskId = req.params.id;
+    const categoryId = req.params.categoryId;
+    const task = await Tasks.findByPk(taskId);
+    if (!task) {
+      return res.status(404).json({ error: 'Task not found' });
+    }
+    const categoryToRemove = await Categories.findByPk(categoryId);
+
+    if (!categoryToRemove) {
+      return res.status(404).json({ error: 'Category not found' });
+    }
+
+    await task.removeCategory(categoryToRemove);
+
+    res.status(200).json({ data: task, categoryToRemove });
+  } catch (error) {
+    next(error);
+    res.status(400).json({ error: 'Delete Category from Task error ' + error });
   }
 };
